Added timer
Timer added using a class in javascript. Code and class are in this same file, we can change this possibly later to have the class in a separate file that can be pulled in. For now, it is easier to see everything together. 
QUESTION: the timer starts on body load, but "0" does not show up. I tried several ways to make it show up, but how can I have it already start with 0 on body load and then count up to 1,2,3, so on?